# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce platform"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Create a new user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["username", "password", "email"]
              }
            }
          },
          "responses": {
            "201": {
              "description": "User created successfully"
            },
            "400": {
              "description": "Bad Request"
            },
            "409": {
              "description": "Username already exists"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Log in a user",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          },
          "responses": {
            "200": {
              "description": "Logged in successfully"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      }
    },
    "/product": {
      "get": {
        "summary": "Get all products",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "List of products"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/product/{productId}": {
      "get": {
        "summary": "Get a product by ID",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "400": {
            "description": "Invalid product ID"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/product/category": {
      "get": {
        "summary": "Get products by category",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products in the category"
          },
          "400": {
            "description": "Missing category ID"
          },
          "404": {
            "description": "No products found for this category"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "summary": "Get a user by ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "400": {
            "description": "Invalid user ID"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{userId}": {
      "put": {
        "summary": "Update a user by ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["username", "email"]
              }
            }
          },
          "responses": {
            "200": {
              "description": "User updated successfully"
            },
            "400": {
              "description": "Invalid user ID"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "/cart": {
      "post": {
        "summary": "Add a product to the cart",
        "tags": ["Cart"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                },
                "required": ["product_id", "quantity"]
              }
            }
          },
          "responses": {
            "201": {
              "description": "Cart created successfully"
            },
            "400": {
              "description": "Please fill all fields"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "/cart/{cartId}": {
      "post": {
        "summary": "Add an item to an existing cart",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                },
                "required": ["product_id", "quantity"]
              }
            }
          },
          "responses": {
            "201": {
              "description": "Item added to cart"
            },
            "404": {
              "description": "Cart not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "/cart/{cartId}": {
      "get": {
        "summary": "Get contents of a specific cart",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart contents retrieved successfully"
          },
          "404": {
            "description": "Cart not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/cart/{cartId}/checkout": {
      "post": {
        "summary": "Checkout the cart",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Checkout successful"
          },
          "400": {
            "description": "Cart is empty"
          },
          "404": {
            "description": "Cart not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Get all orders for the authenticated user",
        "tags": ["Orders"],
        "responses": {
          "200": {
            "description": "List of orders"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "get": {
        "summary": "Get an order by ID",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order found"
          },
          "400": {
            "description": "Invalid order ID"
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartId": {
                    "type": "integer"
                  }
                },
                "required": ["cartId"]
              }
            }
          },
          "responses": {
            "201": {
              "description": "Order created successfully"
            },
            "400": {
              "description": "Invalid cart ID"
            },
            "404": {
              "description": "Cart not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    }
  }
}